refactor(analyzer): move package network_analyzer/http/tools to network_analyzer/tools
