+++ collector/analyzer/tools/http_merger.go
